Add files via upload
The file containing the phrase "Predicting_attributes" has the model for determining the intensity of a fire given that a wildfire is to occur at a given location on a given day. The file containing the phrase "CA_Wildfire_Predictive_Model" is the base model to predict the occurrence of wildfire at a given location (California in this case).